fix: episode file matching issue that double can not convert int. (#368)
* fix: episode file matching issue that double can not convert int.

* docs: update changelog for pr #368.

+++ server/src/main/java/run/ikaros/server/core/subject/service/impl/EpisodeServiceImpl.java
@@ -66,7 +66,7 @@ public class EpisodeServiceImpl implements EpisodeFileService {
             .filter(entity -> FileType.VIDEO.equals(entity.getType()))
             .flatMap(fileEntity -> getSeqMono(fileEntity)
                 .flatMap(seq -> episodeRepository.findBySubjectIdAndSequence(subjectId,
-                    Integer.valueOf(String.valueOf(seq))))
+                    Double.valueOf(String.valueOf(seq))))
                 .flatMap(episodeEntity -> episodeFileRepository
                     .existsByEpisodeId(episodeEntity.getId())
                     .filter(exists -> !exists)
